+++ cursor/decoder.go
@@ -14,12 +14,12 @@ func NewDecoder(model interface{}, keys ...string) (*Decoder, error) {
 	modelType := util.ReflectType(model)
 	// model must be a struct
 	if modelType.Kind() != reflect.Struct {
-		return nil, ErrDecodeInvalidModel
+		return nil, ErrInvalidModel
 	}
 	// validate keys
 	for _, key := range keys {
 		if _, ok := modelType.FieldByName(key); !ok {
-			return nil, ErrDecodeUnknownKey
+			return nil, ErrInvalidModel
 		}
 	}
 	return &Decoder{
@@ -39,23 +39,23 @@ func (d *Decoder) Decode(cursor string) (values []interface{}, err error) {
 	b, err := base64.StdEncoding.DecodeString(cursor)
 	// ensure cursor content is json
 	if err != nil || !json.Valid(b) {
-		return nil, ErrDecodeInvalidCursor
+		return nil, ErrInvalidCursor
 	}
 	jd := json.NewDecoder(bytes.NewBuffer(b))
 	// ensure cursor content is json array
 	if t, err := jd.Token(); err != nil || t != json.Delim('[') {
-		return nil, ErrDecodeInvalidCursor
+		return nil, ErrInvalidCursor
 	}
 	for _, key := range d.keys {
 		// key is already validated when decoder is constructed
 		f, _ := d.modelType.FieldByName(key)
 		v := reflect.New(util.ReflectType(f.Type)).Interface()
 		if err := jd.Decode(&v); err != nil {
-			return nil, ErrDecodeInvalidCursor
+			return nil, ErrInvalidCursor
 		}
 		values = append(values, reflect.ValueOf(v).Elem().Interface())
 	}
 	// cursor must be a valid json after previous checks,
-	// so no need to check "]" is the last token
+	// so no need to check whether "]" is the last token
 	return
 }
