--- cursor/decoder_error.go
@@ -1,10 +1,0 @@
-package cursor
-
-import "errors"
-
-// Errors for decoder
-var (
-	ErrDecodeInvalidCursor = errors.New("invalid cursor for decoding")
-	ErrDecodeInvalidModel  = errors.New("invalid model for decoding")
-	ErrDecodeUnknownKey    = errors.New("unknown key on decoded model")
-)
