--- cursor/encoder_error.go
@@ -1,8 +1,0 @@
-package cursor
-
-import "errors"
-
-// Errors for encoder
-var (
-	ErrEncodeInvalidModel = errors.New("invalid model for encoding")
-)
