+++ cursor/encoder_test.go
@@ -17,12 +17,16 @@ type encoderSuite struct {
 func (s *encoderSuite) TestInvalidModel() {
 	e := NewEncoder("ID")
 	_, err := e.Encode(struct{}{})
-	s.Equal(ErrEncodeInvalidModel, err)
+	s.Equal(ErrInvalidModel, err)
 }
 
 func (s *encoderSuite) TestInvalidModelFieldType() {
 	// https://stackoverflow.com/questions/33903552/what-input-will-cause-golangs-json-marshal-to-return-an-error
 	e := NewEncoder("ID")
-	_, err := e.Encode(struct{ ID chan int }{make(chan int)})
-	s.Equal(ErrEncodeInvalidModel, err)
+	_, err := e.Encode(
+		struct {
+			ID chan int
+		}{make(chan int)},
+	)
+	s.Equal(ErrInvalidModel, err)
 }
