+++ cursor/error.go
@@ -1,0 +1,9 @@
+package cursor
+
+import "errors"
+
+// Errors for encoder
+var (
+	ErrInvalidCursor = errors.New("invalid cursor")
+	ErrInvalidModel  = errors.New("invalid model")
+)
